==> 04_Research_Design_Lifecycle/02_Brainstorming.Rmd <==
---
title: "Idea"
output:
  html_document: default
  pdf_document: default
bibliography: ../bib/book.bib
---

<!-- note do_bookdown is set in index.rmd, so we know if you're running just this .Rmd or compiling the book-->
```{r, eval = !exists("do_bookdown"), echo = FALSE, include = FALSE, purl = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, cache = FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file()) # files are all relative to RStudio project home
```

```{r, eval = !exists("do_bookdown"), echo = FALSE, include = FALSE, purl = FALSE}
# load common packages, set ggplot ddtheme, etc.
source("scripts/before_chapter_script.R")
```

<!-- start post here, do not edit above -->

# Brainstorming

<!-- make sure to rename the section title below -->

```{r idea, echo = FALSE, output = FALSE, purl = FALSE}
# run the diagnosis (set to TRUE) on your computer for this section only before pushing to Github. no diagnosis will ever take place on github.
do_diagnosis <- FALSE
sims <- 100
b_sims <- 20
```

```{r, echo = FALSE}
# load packages for this section here. note many (DD, tidyverse) are already available, see scripts/package-list.R
```

A research design starts from an idea, a kernel of a project. People arrive at their good ideas from every direction through idiosyncratic processes. Some people are inspired by their reading of academic literature. Others are sparked by a conversation with colleagues. Still others have their research questions thrust upon them by exigent circumstances. Ideas come from primary observation of social processes, reading secondary accounts by other authors, and our own past thoughts and experiences. 

Whatever the mysterious process is that ignites the imagination, we can characterize what part or parts of the research design the idea is about. Ideas can be a piece of a model, a vague inquiry, a scrap of a data strategy, or an exciting answer strategy. An idea in the form of a model includes one or more nodes and one or more edges between them: for example, a treatment, a mediator, and an outcome. An inquiry is perhaps the most common kernel of a research project. It might be "does D cause Y." A data strategy might be the discovery of a discontinuity in some administrative rule. A research project that starts with an answer strategy might be a new measurement technology that enables answering previously unanswerable -- or unasked -- questions.

The goal of brainstorming is to build a complete design from this kernel, regardless of where in the four elements *M*, *I*, *D*, or *A* the idea starts out. Sometimes a "theory-first" approach to developing a research project is unhelpfully contrasted with a "methods-first" approach. Better might be to describe the "theory-first" approach as starting with *M* or *I* and an "opportunity-first" approach as starting with *D* or *A*. We think there should be no shame in starting with a data strategy or answer strategy. The ordering is criticized by some for not focusing on "big questions." We think progress in social science can come from many places. We should seek to provide credible answers to important questions, and often progress in providing credible answers will start with new data and answer strategies. That said, unimportant questions aren't worth the research investment, no matter the credibility of the answers -- worse still are unreliable answers to the biggest questions in social science.

## Theory-first: Starting with a model or inquiry

We often come to a research project either with an outcome variable we want to explain or with a question about how a specific treatment affects an outcome. In the terminology of causal graphs, we come with a node (variable) or an edge (the relationship between two variables). 

In this section, we describe in general terms an algorithm for developing theory kernels into research design. The first job is to articulate a complete inquiry. Do we want to know the node's average value, an approximation of its distribution, or its mode? If the kernel is an edge, do we want to understand whether X causes Y, Y causes X, or whether the effects of X on Y interact with a third variable? Defining the inquiry may expand our kernel of an idea: if we started with Y but decide our inquiry is the effect of X on Y, then now we have an edge for our kernel. Next, we want to build out a more complete model around our kernel: what are the possible causes of X and Y, and do they share those causes in common? Since we are almost always quite uncertain about these possible causes, we want to consider a set of alternative possible causal models.

:::: {.principles data-latex=""}
::: {.principles-title data-latex=""}
Algorithm for developing a causal model from a kernel
:::

1. Add common causes of all nodes. Add unknown causes as a node to reflect all of the unknown common causes.
2. Consider adding intermediate variables between connected nodes that represent possible mechanisms.
3. Construct graphs representing all combinations of edges, including lack of edges, between the nodes.
4. Rule out graphs that are cyclic and that are impossible based on substantive knowledge.

::::

The output of this algorithm is not a single DAG, but instead a set of possible DAGs over which we are uncertain. These causal models are nonparametric. In order to encode beliefs about how variables relate to each other or how treatment effects may differ, you may wish to add parametric structure to your set of models. Together, the set models define the model *M* of *MIDA*. With a model and inquiry in hand, we can go in search of a data and answer strategy that together can provide an answer to the inquiry. 

## Opportunity-first: Starting with a data or answer strategy

Research ideas often begin with some snippet of a data strategy: a treatment we can feasibly implement or an opportunity to sample participants from a difficult-to-study population. We can also start the research process with a new answer strategy. Our idea might be a new statistical model or a new way of parsing video data.

We first need to identify an inquiry that might be answered with the new answer strategy. There might be many, but the first step is picking one or a feasibly small number of them. We can then build a model from that inquiry using the algorithm described above. 

In some sense, the opportunity-first approach conditions on feasibility. When an opportunity comes over the transom, we're in the position of taking on the new project or passing. Saying "no" to research opportunities can be difficult. Here our only advice is to imagine that you knew the value of the inquiry with certainty -- what theoretical progress would be achieved? If the answer is "not much," move on.

## Brainstorming sessions 

Brainstorming sessions with colleagues and mentors can help take a kernel of an idea to a set of feasible research designs. What do participants in the brainstorming session need to know about your idea so they can effectively help identify possible designs? We suggest using the following "problem statement.^[The problem statement idea was developed by Graeme Blair, Darin Christensen, Erin Hartman, and Chad Hazlett for a research design seminar at UCLA.]" You need not have all the parts of the design worked out -- that's why you are having a brainstorming session! Systematizing what you do know is helpful. The more information you can provide, even if it is the range of possibilities rather than an exact specification.

:::: {.principles data-latex=""}
::: {.principles-title data-latex=""}
Brainstorming document
:::

Instructions: fill out any part you can!

**Model**
  
- What is the population of units of interest? 
- How many units are in the population? 
- What are the important variables that describe each unit?
- Can you represent your theory as a DAG?
- What parts of the theory are you more or less confident of? 
  
**Inquiry**

- What main question about the theoretical model will the design address?
- Why is this question important for scholarship, the public, or decision-makers?
- Are there auxiliary inquiries that could be used to check model assumptions?

**Data strategy**

- How will you select cases or sample units from the population defined in the model?
- If randomizing treatments:
  - How many conditions will there be?
  - At what level will you randomize?
  - How many units can be assigned to each condition? 
  - Which procedure will you use to randomize?
- If not randomizing treatments:
  - How do units come to receive different treatments?
  - Are there known processes that approximate random assignment
- Regardless of how units come to be treated:
  - How will you address the possibility of noncompliance?
  - How will you address the possibility of spillovers?
  - If there were no financial, logistical, or ethical constraints, what is the ideal experiment you would run?
- How will you measure outcomes?
  - What survey instrument or measurement tool will you use?
  - When will you measure outcomes? How many times?
  - How will you minimize measurement error and attrition?

**Answer strategy**

- How will you use the data that results from your proposed data strategy to produce answers to your inquiry? 
  - What subset of the data will you use analyze?
  - What contrasts will you make to draw comparisons?
  - Which outcomes will you analyze?
  - What estimator will you use?
  - How will you estimate uncertainty in your estimates?

::::
