==> 04_Research_Design_Lifecycle/05_Approvals.Rmd <==
---
title: "Approvals"
output: html_document
bibliography: ../bib/book.bib 
---

<!-- note do_bookdown is set in index.rmd, so we know if you're running just this .Rmd or compiling the book-->
```{r, eval = !exists("do_bookdown"), echo = FALSE, include = FALSE, purl = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, cache = FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file()) # files are all relative to RStudio project home
```

```{r, eval = !exists("do_bookdown"), echo = FALSE, include = FALSE, purl = FALSE}
# load common packages, set ggplot ddtheme, etc.
source("scripts/before_chapter_script.R")
```

<!-- start post here, do not edit above -->

## Approvals

<!-- make sure to rename the section title below -->

```{r approvals, echo = FALSE, output = FALSE, purl = FALSE}
# run the diagnosis (set to TRUE) on your computer for this section only before pushing to Github. no diagnosis will ever take place on github.
do_diagnosis <- FALSE
sims <- 100
b_sims <- 20
```

```{r, echo = FALSE}
# load packages for this section here. note many (DD, tidyverse) are already available, see scripts/package-list.R
```

When researchers sit at universities in the United States, research must be approved by the university's institutional review board (IRB) under the federal regulation known as the "Common Rule." Similar research review bodies exist at universities worldwide and at many independent research organizations and think tanks. Though these boards are commonly thought to judge research ethics, in fact, they mainly exist to protect their institution from liability for research gone awry [@King2015]. Accordingly, a researcher's obligation to consider their study's ethics is neither constrained nor checked by IRBs. Instead, a set of idiosyncratic rules and practices specific to each institution are checked. The researcher, as a result, remains responsible for their own ethical decision about whether or not to move forward with the research. That said, the IRB process is not necessarily without benefit. In some cases, useful discussions can be had with IRB board members about study decisions, and the approval itself may protect the researcher from some kinds of liability. 

Laws and regulations at the country, state or province, or municipality level may also govern research on human subjects besides the IRB. Many countries require human subjects approval, especially for health research, in addition to the approvals researchers must seek from their home institutions. These approvals serve a similar purpose to the home institution IRB, but by virtue of their authority coming from the context in which the research is conducted rather than from far away bureaucrats, they may serve to more directly protect human subjects.

Though these bodies' goals differ from the broader ethical aims social scientists hold, design diagnosis may also be useful here. Many IRBs ask researchers to describe tradeoffs between the costs and benefits to research subjects. In some cases, researchers are asked to defend research design choices that provide benefits to science, but where the only direct effects on participants are costs with no immediate benefits. Defining the costs and benefits to participants in terms of their time and money and the compensation provided by researchers, if any, can both simplify communication with IRBs and provide tools for researchers to more easily clarify these tradeoffs for themselves. The expected benefit and expected cost can be diagnosands across possible realizations of the design. The design diagnosis can highlight tradeoffs between the value to participants and the scientific value in the form of standard diagnosands. Rather than argue in the abstract about these quantities, they can be simulated and described formally through declaration and diagnosis.

**Further readings.**

- [U.S. "Revised Common Rule"](https://www.hhs.gov/ohrp/regulations-and-policy/regulations/finalized-revisions-common-rule/index.html)

<!-- Local research approval -->

<!-- Partner approval  -->


<!-- - IRB approval (purpose is to protect the university from liability) - Gary and Melissa's paper -->
<!-- - IRB approval at your institution and with others -->
<!-- - IRB in context where you are working - country IRBs, what to do if there is not a country IRB -->
<!-- - research clearances -->


<!-- - history: belmont report. tuskegee experiment. stanford prison experiment.  -->
<!-- - laws and IRBs: revised common rule, paper on IRBs -->
<!-- - guidelines from econ, poli sci, soc, psych. -->
