==> 04_Research_Design_Lifecycle/06_Partners.Rmd <==
---
title: "Partners"
output: html_document
bibliography: ../bib/book.bib 
---

<!-- note do_bookdown is set in index.rmd, so we know if you're running just this .Rmd or compiling the book-->
```{r, eval = !exists("do_bookdown"), echo = FALSE, include = FALSE, purl = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, cache = FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file()) # files are all relative to RStudio project home
```

```{r, eval = !exists("do_bookdown"), echo = FALSE, include = FALSE, purl = FALSE}
# load common packages, set ggplot ddtheme, etc.
source("scripts/before_chapter_script.R")
```

<!-- start post here, do not edit above -->

## Partners

<!-- make sure to rename the section title below -->

```{r partners, echo = FALSE, output = FALSE, purl = FALSE}
# run the diagnosis (set to TRUE) on your computer for this section only before pushing to Github. no diagnosis will ever take place on github.
do_diagnosis <- FALSE
sims <- 100
b_sims <- 20
```

```{r, echo = FALSE}
# load packages for this section here. note many (DD, tidyverse) are already available, see scripts/package-list.R
```

Partnering with third-party organizations in research entails cooperating to intervene in the world or to measure outcomes. Researchers seek to produce (and publish) scientific knowledge; they work with political parties, government agencies, nonprofit organizations, and businesses to learn more than they could if they worked independently. These groups work with researchers to learn about how to achieve their own organizational goals. Governments may want to expand access to healthcare, corporations to improve their ad targeting, and nonprofits to demonstrate program impact to funding organizations.

In the best-case scenario, the goals of the researchers and partner organizations are aligned. When the scientific question to be answered is the same as the practical question the organization cares about, the gains from cooperation are clear. The research team gains access to the organization's financial and logistical capacity to act in the world, and the partner organization gains access to the researchers' scientific expertise. Finding the right research partner almost always amounts to finding an organization with a common -- or at least not conflicting -- goal. Selecting a research design amenable to both parties requires understanding each partners' private goals. Research design declaration and diagnosis can help with this problem by formalizing tradeoffs between the two sets of goals.

One frequent divergence between partner and researcher goals is that partner organizations often want to learn, but they care most about their primary mission. This dynamic is sometimes referred to as the "learning versus doing" tradeoff. (In business settings, this tradeoff goes by names like "learning versus earning" or "exploration versus exploitation"). An aid organization cares about delivering their program to as many people as possible. Learning whether the program has the intended effects on the outcomes of interest is obviously also important, but resources spent on evaluation are resources *not* spent on program delivery. 

Research design diagnosis can help navigate the learning versus doing tradeoff. One instance of the tradeoff is that the proportion of units that receive a treatment represents the rate of "doing," but this rate also affects the amount of learning. In the extreme, if all units are treated, we can't measure the effect of the treatment. The tradeoff here is represented in a graph of the study's power versus the proportion treated (top facet) and the partner's utility (bottom facet). The researchers have a power cutoff at the standard 80% threshold. The partner also has a strict cutoff: they need to treat at least 2/3 of the sample to fulfill a donor requirement. 

Researchers might simply ignore the proportion treated and select the design with the highest power in the absence of partners. With a partner organization, the researcher might use this graph in conversation with the partner to jointly select the design that has the highest power that has a sufficiently high proportion treated to meet the partner's needs. This is represented in the "zone of agreement" in gray: in this region, the design has at least 80% power and at least two-thirds of the sample are treated. Deciding within this region involves a tradeoff between power (which is decreasing in the proportion treated here) and the partner's utility (which is increasing in proportion treated). The diagnosis surfaces the zone of the agreement and clarifies the choice between designs in that region.^[Unfortunately, some partnerships simply will not work out if the zone of agreement is empty.]

```{r, echo = FALSE, eval = do_diagnosis & !exists("do_bookdown")}
design <-
  declare_model(N = 100,
                U = rnorm(N, sd = 0.2),
                potential_outcomes(Y ~ 0.15 * Z + U)) +
  declare_assignment(Z = complete_ra(N = N, prob = prob), legacy = FALSE) +
  declare_measurement(Y = reveal_outcomes(Y ~ Z)) +
  declare_estimator(Y ~ Z)

designs <- redesign(design, prob = seq(1/12, 11/12, length.out = 11))
  
simulations <- simulate_designs(designs, sims = 1000)

gg_df <-
  simulations %>%
  group_by(prob) %>% 
  summarise(Power = mean(p.value <= 0.05),
            Utility = prob,
            acceptable_power = Power >= 0.80,
            acceptable_utility = Utility >= 2/3, .groups = "drop") %>% 
  pivot_longer(c(Power, Utility),names_to = "diagnosand")
```

```{r, echo = FALSE, purl = FALSE}
# figure out where the dropbox path is, create the directory if it doesn't exist, and name the RDS file
rds_file_path <- paste0(get_dropbox_path("04_Partners.Rmd"), "/gg_df.RDS")
if (do_diagnosis & !exists("do_bookdown")) {
  write_rds(gg_df, file = rds_file_path)
}
gg_df <- read_rds(rds_file_path)
```

```{r partnersplot, echo = FALSE, fig.cap = "Navigating research partnerships.", fig.width = 6.5, fig.height = 6.5}
rect_df <- 
  gg_df %>%
  filter(acceptable_power, acceptable_utility) %>%
  summarise(min_prob = min(prob),
            max_prob = max(prob))

lines_df <-
  tibble(
    diagnosand = c("Power", "Utility"),
    threshold = c(0.8, 2/3)
  )

annotations_df <-
  tibble(
    diagnosand = c("Power", "Utility"),
    label = c("Researcher constraint: Power must be above 0.80.",
              "Partner constraint: At least two-thirds of subjects must be treated."),
    prob = c(0.2, 0.1),
    value = c(0.725, 0.725)
  )

ggplot(gg_df, aes(prob, value)) +
  geom_point() +
  geom_line(alpha = 0.6) +
  geom_text(data = annotations_df,
            aes(label = label),
            hjust = 0) +
  geom_rect(
    data = rect_df,
    aes(
      xmin = min_prob,
      xmax = max_prob,
      ymin = 0,
      ymax = 1
    ),
    alpha = 0.1,
    inherit.aes = FALSE
  ) +
  geom_hline(data = lines_df, aes(yintercept = threshold)) +
  annotate(
    "text",
    x = 17/24,
    y = 0.60,
    label = "Zone of agreement",
    angle = 270,
    hjust = 0
  ) +
  scale_y_continuous(breaks = seq(0, 1, length.out = 6)) +
  scale_x_continuous(breaks = seq(0, 1, length.out = 7),
                     labels = c("0/6", "1/6", "2/6", "3/6", "4/6", "5/6", "6/6")) +
  facet_grid(diagnosand ~ ., scales = "free") +
  labs(x = "Proportion treated") +
  dd_theme() +
  theme(axis.title.y = element_blank(),
        panel.grid.minor.y = element_blank())
```

Choosing the proportion treated is one example of integrating partner constraints into research designs. A second common problem is that there are a set of units that must be treated or that must not be treated for ethical or political reasons (e.g., the home district of a government partner must receive the treatment). If these constraints are discovered after treatment assignment, they lead to noncompliance, which may substantially complicate the analysis of the experiment and even prevent providing an answer to the original inquiry. @Gerber2012 recommend, before randomizing treatment, exploring possible treatment assignments with the partner organization and using this exercise to elicit the set of units that must or cannot be treated. @king2007politically describe a "politically-robust" design, which uses pair-matched block randomization. In this design, when any unit is dropped due to political constraints, the whole pair is dropped from the study.^[This procedure is prone to bias for the average treatment effect among the "political feasible" units if within some pairs, one unit is treatable but the other is not.]

A major benefit of working with partners is their deep knowledge of the substantive area. For this reason, we recommend involving them in the design declaration and diagnosis process. How can we develop intuitions about the means, variances, and covariances of the variables to be measured? Ask your partner for their best guesses, which may be far more educated than your own. For experimental studies, solicit your partner's beliefs about the magnitude of the treatment effect on each outcome variable, subgroup by subgroup if possible. Engaging partners in the declaration process improves design -- and it very quickly sharpens the discussion of key design details. Sharing your design diagnoses and mock analyses *before* the study is launched can help to build a consensus around the study's goals.


**Further Reading**

- @king_persily_2020
- @levi_rajala_2020
- @levine_2020
- @levine_2020b
- [JPAL research resources: "Formalize research partnership and establish roles and expectations"](https://www.povertyactionlab.org/resource/formalize-research-partnership-and-establish-roles-and-expectations)
- [JPAL research resources: "Assessing viability and building relationships"](https://www.povertyactionlab.org/resource/assessing-viability-and-building-relationships)

<!-- ### Scattered thoughts -->

<!-- - Partnerships and ethics should be in here -->
<!-- - Research for Impact (@levine_2020) -->
<!-- - When to walk away -->
<!-- - Advice: there's usually a boss that gives you the green light, then staff people who actually help you. Our advice is to ensure that at least some meetings are with both the boss and the staff people so that the staff people know that the boss cares about the project and are therefore motivated to help you. A crucial person at the partner organization is the person who knows where the data are and how you can access them. Many people at partner organizations do not have this power, but you must be frequent touch.  -->
<!-- - In some cases, some member of the research staff can actually embed wiht the partner organization to ensure all the good things. -->
<!-- - Involve partners in PAP writing. -->
<!-- - when should partners be coauthors on the paper? -->




<!-- (Increasingly, funders are separating evaluation funds from implementation funds to remove this tradeoff.) -->
