==> 04_Research_Design_Lifecycle/14_Reconciliation.Rmd <==
---
title: "Reconciliation"
output: html_document
bibliography: ../bib/book.bib 
---

<!-- note do_bookdown is set in index.rmd, so we know if you're running just this .Rmd or compiling the book-->
```{r, eval = !exists("do_bookdown"), echo = FALSE, include = FALSE, purl = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, cache = FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file()) # files are all relative to RStudio project home
```

```{r, eval = !exists("do_bookdown"), echo = FALSE, include = FALSE, purl = FALSE}
# load common packages, set ggplot ddtheme, etc.
source("scripts/before_chapter_script.R")
```

<!-- start post here, do not edit above -->

## Reconciliation {#p4reconciliation}

<!-- make sure to rename the section title below -->

```{r reconciliation, echo = FALSE, output = FALSE, purl = FALSE}
# run the diagnosis (set to TRUE) on your computer for this section only before pushing to Github. no diagnosis will ever take place on github.
do_diagnosis <- FALSE
sims <- 100
b_sims <- 20
```

```{r, echo = FALSE}
# load packages for this section here. note many (DD, tidyverse) are already available, see scripts/package-list.R
```

Research design as implemented will differ in some way from research designs as planned. Treatments cannot be executed as conceived, some people cannot be found to interview, and sometimes what we learn from baseline measures informs how we measure later. Understanding how your research design changed from conception to implementation is crucial to understanding what was learned from the study.

Suppose the original design described a three-arm trial: one control and two treatments, but the design as implemented drops all subjects assigned to the second treatment. Sometimes, this is an entirely appropriate and reasonable design modification. Perhaps it turns out that the second treatment was simply not delivered due to an implementation failure. Other times, these modifications are less benign: perhaps the second treatment effect estimate does not achieve statistical significance, so the author omits it from the analysis. 

For this reason, we recommend that authors reconcile the design as planned with the design as implemented. A reconciliation can be a plain description of the deviations from the PAP, with justifications where appropriate. A more involved reconciliation would include a declaration of the planned design, a declaration of the implemented design, and a list of the differences. This "diff" of the designs can be automated through the declaration of both designs in computer code, then comparing the two design objects line-by-line (see the function `compare_designs()` in DeclareDesign). 

In some cases, reconciliation will lead to additional learning beyond what can be inferred from the final design itself. When some units refuse to be included in the study sample or some units refuse measurement, we learn that important imformation about those units. Understanding sample exclusions, noncompliance, and attrition not only may inform future research design planning choices but contribute substantively to our understanding of the social setting.

### Example

In Section \@ref(p4planning), we described the preanalysis plan registered by @bonilla_tillery_2020. We reconcile the set of conditional average treatment effect (CATE) analyses planned in that PAP, the analyses reported in the paper, and those reported in the appendix at the request of reviewers in Table \@ref(tab:reconciliation). In column two, we see that the authors planned four CATE estimations: effects by familiarity with Black Lives Matter; by gender; LGBTQ status; and linked fate. Only two of those are reported in the paper; the others may have been excluded for space reasons. Another way to handle these uninteresting results would be to present them in a populated PAP posted on their Web site or in the paper's appendix.

In their appendix, the authors report on a set of analyses requested by reviewers. We see this as a perfect example of transparently presenting the set of planned analyses and highlighting the analyses that were added afterward and why they were added. They write:

> We have been asked to consider other pertinent moderations beyond gender and LGBTQ+ status. They are contained in the four following sections.

This small table describes the heterogeneous effects analyses the researchers planned, those reported in the paper, and those reported in the appendix at the request of reviewers.


| Covariate            | In the preanalysis plan | In the paper            | In the appendix (at the request of reviewers) |
| -------------------- | ----------------------- | ----------------------- | --------------------------------------------- |
| Familiarity with BLM | $\checkmark$            |                         |                                               |
| Gender               | $\checkmark$            | $\checkmark$            |                                               |
| LGBTQ status        | $\checkmark$            | $\checkmark$            |                                               |
| Linked fate          | $\checkmark$            |                         |                                               |
| Religiosity          |                         |                         | $\checkmark$                                  |
| Region               |                         |                         | $\checkmark$                                  |
| Age                  |                         |                         | $\checkmark$                                  |
| Education            |                         |                         | $\checkmark$                                  |

Table: (\#tab:reconciliation) Reconciliation of Bonilla and Tillery preanalysis plan.


<!-- ### Scattered thoughts -->

<!-- - @ofosu2019pre -->



<!-- Having a publicly-posted preanalysis plan can make the reconciliation process especially credible -- we know for sure what the planned design was because the preanalysis plan describes it pre-implementation. However, a preanalysis plan is not a prerequisite for engaging in reconciliation. The scientific enterprise is built in large measure on trust: we are ready to believe researchers who say, here is the design I thought I would implement but due to unanticipated developments, here is the design I ended up implementing.  -->
